+++ Functions.h
@@ -13,11 +13,11 @@ using namespace std;
 *
 *******************************************************/
 
-bool isQuadratic(const long& n, const long& p);
-bool isPrime(long& n);
-bool isRoot(long& a, long& p);
-bool isHere(vector<long> differences, long diff);
-bool isCostas(vector<long>& shift);
+bool isQuadratic(const int& n, const int& p);
+bool isPrime(int& n);
+bool isRoot(int& a, int& p);
+bool isHere(vector<int> differences, int diff);
+bool isCostas(vector<int>& shift);
 
 
 /******************************************************
@@ -26,11 +26,11 @@ bool isCostas(vector<long>& shift);
 *
 *******************************************************/
 
-unsigned long readDim();
-bool compare(vector<long>& v1, vector<long>& v2);
-long powerMod(const long& x, const long& n, const long& mod);
-long f(const long& x, const long& mod, long& a, long& b);
-long LegendreComplexity(long& p);
+unsigned int readDim();
+bool compare(vector<int>& v1, vector<int>& v2);
+int powerMod(const int& x, const int& n, const int& mod);
+int f(const int& x, const int& mod, int& a, int& b);
+int LegendreComplexity(int& p);
 
 
 /******************************************************
@@ -41,21 +41,21 @@ long LegendreComplexity(long& p);
 
 
 struct MSeq {
-    vector<long> v;
+    vector<int> v;
     MSeq() {
         v = {0,0,1,0,1,1,1};
     }
-    NTL::ZZ_p operator () (const long& i) const {
+    NTL::ZZ_p operator () (const int& i) const {
         return (NTL::ZZ_p) v[i];
     }
 };
 
 struct LegendreSeq {
-    long mod;
-    explicit LegendreSeq(long p) {
+    int mod;
+    explicit LegendreSeq(int p) {
         mod = p;
     }
-    NTL::ZZ_p operator () (const long& i) const {
+    NTL::ZZ_p operator () (const int& i) const {
         bool result;
         if ((i % mod) == 0)
             result = false;
@@ -66,99 +66,99 @@ struct LegendreSeq {
 };
 
 struct PolynomialSeq {
-    long m;
-    long alpha;
-    long A;
-    long B;
-    explicit PolynomialSeq(long mod, long root, long a, long b) {
+    int m;
+    int alpha;
+    int A;
+    int B;
+    explicit PolynomialSeq(int mod, int root, int a, int b) {
         m = mod; alpha = root; A = a; B = b;
     }
-    long operator () (const long& i) {
+    int operator () (const int& i) {
         return f(powerMod(alpha, i, m), m, A, B);
     }
 };
 
 
 struct ExpQuadraticSeq {
-    long mod;
-    long root;
-    long a;
-    explicit ExpQuadraticSeq(long p, long r, long A) {
+    int mod;
+    int root;
+    int a;
+    explicit ExpQuadraticSeq(int p, int r, int A) {
         mod = p; root = r; a = A;
     }
-    long operator () (const long& i) {
+    int operator () (const int& i) {
         return (a * powerMod(root*root, i, mod) + powerMod(root, i, mod)) % mod;
     }
 };
 
 struct LogQuadraticSeq {
-    long mod;
-    long root;
-    vector<long> logTable;
-    explicit LogQuadraticSeq(long p, long r) {
+    int mod;
+    int root;
+    vector<int> logTable;
+    explicit LogQuadraticSeq(int p, int r) {
         mod = p; root = r;
         logTable.resize(p);
         logTable[0] = -1;
-        for (long i=0; i<mod-1; i++) {
+        for (int i=0; i<mod-1; i++) {
             logTable[powerMod(root, i, mod)] = i;
         }
     }
-    long operator () (const long& i) {
-        long result = (4*powerMod(root*root, i, mod) + powerMod(root, i, mod)) % mod;
+    int operator () (const int& i) {
+        int result = (4*powerMod(root*root, i, mod) + powerMod(root, i, mod)) % mod;
         return logTable[result];
     }
 };
 
 
 struct ExponentialSeq {
-    long mod;
-    long root;
-    explicit ExponentialSeq(long m, long r){
+    int mod;
+    int root;
+    explicit ExponentialSeq(int m, int r){
         mod = m; root = r;
     }
-    long operator () (const long& i) const {
+    int operator () (const int& i) const {
         return powerMod(root, i, mod);
     }
 };
 
 struct OneSequence {
-    long n;
-    explicit OneSequence(long length) {
+    int n;
+    explicit OneSequence(int length) {
         n = length;
     }
-    NTL::ZZ_p operator () (const long& i) const {
+    NTL::ZZ_p operator () (const int& i) const {
         if(i==0)    return (NTL::ZZ_p) 1;
         else        return (NTL::ZZ_p) 0;
     }
 };
 
 struct LosWelchSeq {
-    long mod;
-    long root;
-    vector<long> log;
-    explicit LosWelchSeq(long m, long r){
+    int mod;
+    int root;
+    vector<int> log;
+    explicit LosWelchSeq(int m, int r){
         mod = m;
         root = r;
         log.resize(m);
-        for (long i=0; i<m-1; i++){
+        for (int i=0; i<m-1; i++){
             log[powerMod(r,i,m)-1] = i;
         }
     }
-    long operator () (const long& i) const {
+    int operator () (const int& i) const {
         return log[i];
     }
 };
 
 
 template <typename Func>
-void printSeq(Func func1, long period){
-    for (long i = 0; i < period; i++) {
+void printSeq(Func func1, int period){
+    for (int i = 0; i < period; i++) {
         cout << func1(i) << " ";
     }
     cout << endl;
 }
 
-long LogRootSeq(long i);
-long noShiftFunc(const long& i);
+int LogRootSeq(int i);
+int noShiftFunc(const int& i);
 
 #endif //RST_FUNCTIONS_H
